+++ lib/vagrant/machine.rb
@@ -211,7 +211,11 @@ module Vagrant
         # Delete the entire data directory contents since all state
         # associated with the VM is now gone.
         @data_dir.children.each do |child|
-          child.rmtree
+          begin
+            child.rmtree
+          rescue Errno::EACCES
+            @logger.info("EACCESS deleting file: #{child}")
+          end
         end
       end